refactor: rename device factory class name
